+++ R/feature_groups-screening.R
@@ -639,7 +639,7 @@ genIDLevelRulesFile <- function(out, inLevels = NULL, exLevels = NULL)
         if (!is.null(exLevels))
             rules <- rules[!grepl(exLevels, names(rules))]
         # UNDONE: this quotes ID levels without sub-level, fix?
-        yaml::write_yaml(rules, out, indent = 4)
+        writeYAML(rules, out)
     }
     invisible(NULL)
 }